CT-FM/requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
astroid==3.3.8
attrs==25.1.0
babel==2.17.0
cerberus==1.3.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dill==0.3.9
docker-pycreds==0.4.0
filelock==3.17.0
fire==0.5.0
frozenlist==1.5.0
fsspec==2025.2.0
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.44
griffe==1.5.6
grpcio==1.70.0
hydra-core==1.4.0.dev1
idna==3.10
importlib-resources==6.5.2
isort==6.0.0
itk==5.4.0
itk-core==5.4.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
jinja2==3.1.5
lighter==0.0.3a18
lightly==1.5.18
lightly-utils==0.0.2
lightning-utilities==0.12.0
loguru==0.6.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.3.0
mkdocs-gen-files==0.5.0
mkdocs-get-deps==0.2.0
mkdocs-literate-nav==0.6.1
mkdocs-material==9.6.2
mkdocs-material-extensions==1.3.1
mkdocs-section-index==0.3.9
mkdocstrings==0.28.0
mkdocstrings-python==1.14.5
monai==1.4.1rc1
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.4.0.dev3
packaging==24.2
paginate==0.5.7
pandas==1.5.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pydantic==2.11.0a1
pydantic-core==2.28.0
pygments==2.19.1
pylint==3.3.4
pymdown-extensions==10.14.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
pyyaml==6.0.2
pyyaml-env-tag==0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tomlkit==0.13.2
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing-extensions==4.12.2
urllib3==2.3.0
wandb==0.19.6
watchdog==6.0.0
werkzeug==3.1.3
yarl==1.18.3
